requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pydantic>=2.0.0
httpx>=0.27.0
